Revision from Christines feedback
Hi Nixon,

Thank you for sharing your draft.

Please see below for feedback to consider:

Remove the social links in the footer – or make them link to something relevant.

It is difficult to read the word ‘Menu’ on mobile. Perhaps change the font on this breakpoint to something else.
A black and white photo

Description automatically generated

Consider adding some margin above the subtitle on the Robin Hood, Kelpies, King Arthur, and Lady Godiva pages. It is quite close to the in-page navigation, particularly on mobile.

A close up of a calendar

Description automatically generated

A screenshot of a phone

Description automatically generated

Give some context to the videos about the Kelpie and Lady Godiva. Why have you included it on the page? Is it something you’re recommending to your visitors to watch and why?

On the King Arthur page, the images justify to the left on mobile. Consider if these could work centrally aligned.

A person with a beard and a helmet

Description automatically generated

Take care with image file size. I’ve looked in your GitHub repository and some are still in MBs.

You have made a good start with your design process document. Make sure to include annotations under all screenshots. For example, in your screenshots about the Trello board, make sure to add brief notes about your thoughts on the Kanban method and your general progress.

You can include one end-of-text reference list at the end of document, rather than reference lists after each individual section.

Say I, not we – as this is your project and own work.

I hope the above is helpful.

Keep up the hard work,
Many thanks,

Christine